bike_store/requirements.txt
appnope==0.1.0
backcall==0.1.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
Django==2.1.3
django-rest-framework==0.1.0
djangorestframework==3.9.0
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-datastore==1.7.1
googleapis-common-protos==1.5.5
grpcio==1.16.1
gunicorn==19.9.0
httplib2==0.12.0
idna==2.7
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
oauth2client==4.1.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.3.0
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
